lots of new filtering functions
on the item-manager side, you should be able to filter anything in the requirements. Choose what to filter by passing in values to setters.

This is missing a lot of things like an interface, and connection to the page where you choose how to sort. I also require items to keep track of what tags they have, and that is not present.